requirements.txt
pynput == 1.6.8
pyyaml == 5.3.1